Move offline account store and bat cache to client package
