+++ src/peergos/shared/login/LoginCache.java
@@ -1,0 +1,13 @@
+package peergos.shared.login;
+
+import peergos.shared.crypto.asymmetric.*;
+import peergos.shared.user.*;
+
+import java.util.concurrent.*;
+
+public interface LoginCache {
+
+    CompletableFuture<Boolean> setLoginData(LoginData login);
+
+    CompletableFuture<UserStaticData> getEntryData(String username, PublicSigningKey authorisedReader);
+}
